feat: audio downloading

+++ src/commands/downloader.ts
@@ -7,11 +7,11 @@ import * as dotenv from 'dotenv';
 dotenv.config();
 
 const RAPIDAPI_KEY = process.env.RAPIDAPI_KEY;
-const FILE_SIZE_LIMIT_MB = 50;  // Telegram fayl hajmi cheklovi
+const FILE_SIZE_LIMIT_MB = 50;  // Telegram file size limit
 const MAX_RETRIES = 3;  // Retry limit
 
 export const handleDownloadCommand = (bot: TelegramBot, chatId: number) => {
-    bot.sendMessage(chatId, 'Iltimos, yuklab olmoqchi bo\'lgan video yoki rasmning URL manzilini yuboring.');
+    bot.sendMessage(chatId, 'Please send the URL of the video or image you want to download.');
 };
 
 const downloadVideo = async (url: string, output: string): Promise<void> => {
@@ -28,17 +28,27 @@ const downloadVideo = async (url: string, output: string): Promise<void> => {
     });
 };
 
-const getVideoSizeInMB = (filePath: string): number => {
+const downloadAudio = async (url: string, output: string): Promise<void> => {
+    const response = await axios({
+        method: 'GET',
+        url: url,
+        responseType: 'arraybuffer',
+    });
+
+    return fs.promises.writeFile(output, response.data);
+};
+
+const getFileSizeInMB = (filePath: string): number => {
     const stats = fs.statSync(filePath);
     const fileSizeInBytes = stats.size;
     return fileSizeInBytes / (1024 * 1024);  // Convert to MB
 };
 
-const fetchFromApi = async (url: string, apiUrl: string, host: string, retries: number = MAX_RETRIES): Promise<any> => {
+const fetchFromApi = async (params: object, apiUrl: string, host: string, retries: number = MAX_RETRIES): Promise<any> => {
     for (let attempt = 0; attempt < retries; attempt++) {
         try {
             return await axios.get(apiUrl, {
-                params: { url: url },
+                params: params,
                 headers: {
                     'X-RapidAPI-Key': RAPIDAPI_KEY,
                     'X-RapidAPI-Host': host
@@ -59,29 +69,32 @@ export const handleMediaUrl = async (bot: TelegramBot, msg: TelegramBot.Message)
     const url = msg.text;
 
     if (url && url.startsWith("http")) {
-        bot.sendMessage(chatId, 'Videoni yuklab olmoqda, biroz kuting...');
+        bot.sendMessage(chatId, 'Processing the URL, please wait...');
 
         try {
             let downloadUrl: string | null = null;
             let title: string | null = null;
+            let isVideo = false;
 
             if (url.includes("youtube.com") || url.includes("youtu.be")) {
                 const videoId = extractVideoId(url);
                 console.log('videoId:', videoId);
                 if (!videoId) {
-                    throw new Error('Yaroqsiz YouTube URL. Iltimos, to\'g\'ri URL kiriting.');
+                    throw new Error('Invalid YouTube URL. Please enter a correct URL.');
                 }
 
-                const apiResponse = await fetchFromApi(url, 'https://yt-api.p.rapidapi.com/dl', 'yt-api.p.rapidapi.com');
+                let apiResponse;
+                apiResponse = await fetchFromApi({ videoId: videoId }, 'https://youtube-media-downloader.p.rapidapi.com/v2/video/details', 'youtube-media-downloader.p.rapidapi.com');
 
-                console.log('API Response:', apiResponse.data);  // Javobni konsolga chiqarish
+                console.log('API Response:', apiResponse.data);  // Log the response
 
-                if (apiResponse.data.status === 'fail') {
-                    throw new Error(apiResponse.data.error);
+                if (!apiResponse.data.status || apiResponse.data.errorId !== 'Success') {
+                    throw new Error('Invalid API response');
                 }
 
-                downloadUrl = apiResponse.data.downloadUrl || apiResponse.data.formats?.[0]?.url;
+                downloadUrl = apiResponse.data.videos.items[0].url;
                 title = apiResponse.data.title;
+                isVideo = true;
 
                 if (!downloadUrl || !title) {
                     throw new Error('Invalid API response');
@@ -91,13 +104,13 @@ export const handleMediaUrl = async (bot: TelegramBot, msg: TelegramBot.Message)
                 let apiResponse;
 
                 try {
-                    apiResponse = await fetchFromApi(url, 'https://instagram-downloader.p.rapidapi.com/index', 'instagram-downloader.p.rapidapi.com');
+                    apiResponse = await fetchFromApi({ url: url }, 'https://instagram-downloader.p.rapidapi.com/index', 'instagram-downloader.p.rapidapi.com');
                 } catch (error) {
                     console.error('Primary API failed:', error);
-                    apiResponse = await fetchFromApi(url, 'https://instagram-downloader-download-photo-video-reels-igtv.p.rapidapi.com/data', 'instagram-downloader-download-photo-video-reels-igtv.p.rapidapi.com');
+                    apiResponse = await fetchFromApi({ url: url }, 'https://instagram-downloader-download-photo-video-reels-igtv.p.rapidapi.com/data', 'instagram-downloader-download-photo-video-reels-igtv.p.rapidapi.com');
                 }
 
-                console.log('API Response:', apiResponse.data);  // Javobni konsolga chiqarish
+                console.log('API Response:', apiResponse.data);  // Log the response
 
                 if (isStory) {
                     if (!apiResponse.data.result || !apiResponse.data.result.video_url) {
@@ -105,42 +118,86 @@ export const handleMediaUrl = async (bot: TelegramBot, msg: TelegramBot.Message)
                     }
                     downloadUrl = apiResponse.data.result.video_url;
                     title = apiResponse.data.result.username || 'instagram_story';
+                    isVideo = true;
                 } else {
                     if (!apiResponse.data.result || !apiResponse.data.result.video_url) {
                         throw new Error('Invalid API response');
                     }
                     downloadUrl = apiResponse.data.result.video_url;
                     title = apiResponse.data.result.username || 'instagram_video';
+                    isVideo = true;
                 }
 
                 if (!downloadUrl || !title) {
                     throw new Error('Invalid API response');
                 }
+            } else if (url.match(/\.(jpeg|jpg|gif|png)$/) != null) {
+                await bot.sendPhoto(chatId, url);
+                return;
             } else {
-                throw new Error('Yaroqsiz URL. Iltimos, to\'g\'ri URL kiriting.');
+                throw new Error('Invalid URL. Please enter a correct URL.');
             }
 
             const sanitizedTitle = title.replace(/[^a-zA-Z0-9]/g, '_');
-            const output = path.resolve(__dirname, `${sanitizedTitle}.mp4`);
 
-            await downloadVideo(downloadUrl, output);
+            if (isVideo) {
+                await bot.sendMessage(chatId, 'Do you want the video or just the audio (MP3)?', {
+                    reply_markup: {
+                        inline_keyboard: [
+                            [
+                                { text: 'Video', callback_data: 'video' },
+                                { text: 'Audio', callback_data: 'audio' }
+                            ]
+                        ]
+                    }
+                });
 
-            const fileSizeMB = getVideoSizeInMB(output);
-            if (fileSizeMB > FILE_SIZE_LIMIT_MB) {
-                await bot.sendMessage(chatId, `Video hajmi ${FILE_SIZE_LIMIT_MB} MB dan katta. Iltimos, boshqa video kiriting.`);
-            } else {
-                const botUsername = 'tg_multitask_bot';  // Bot username'ini o'zgartiring yoki dinamik tarzda oling
-                await bot.sendMessage(chatId, "Yuklash yakunlandi. Videoni yubormoqda...");
-                await bot.sendDocument(chatId, output, { caption: `Video yuklandi. @${botUsername}` });
+                bot.once('callback_query', async (callbackQuery) => {
+                    const action = callbackQuery.data;
+                    const userId = callbackQuery.from.id;
+                    const replyMessageId = callbackQuery.message?.message_id;
+
+                    if (replyMessageId) {
+                        await bot.deleteMessage(userId, replyMessageId.toString());
+                    }
+
+                    if (action === 'video') {
+                        const output = path.resolve(__dirname, `${sanitizedTitle}.mp4`);
+
+                        await downloadVideo(downloadUrl, output);
+
+                        const fileSizeMB = getFileSizeInMB(output);
+                        if (fileSizeMB > FILE_SIZE_LIMIT_MB) {
+                            await bot.sendMessage(chatId, `The video size is larger than ${FILE_SIZE_LIMIT_MB} MB. Please upload another video.`);
+                        } else {
+                            await bot.sendChatAction(chatId, 'upload_video');
+                            await bot.sendVideo(chatId, output, { caption: `Video uploaded. @tg_multitask_bot` });
+                        }
+
+                        fs.unlinkSync(output);  // Delete the file from the server
+                    } else if (action === 'audio') {
+                        const output = path.resolve(__dirname, `${sanitizedTitle}.mp3`);
+
+                        await downloadAudio(downloadUrl, output);
+
+                        const fileSizeMB = getFileSizeInMB(output);
+                        if (fileSizeMB > FILE_SIZE_LIMIT_MB) {
+                            await bot.sendMessage(chatId, `The audio size is larger than ${FILE_SIZE_LIMIT_MB} MB. Please upload another audio.`);
+                        } else {
+                            await bot.sendChatAction(chatId, 'upload_audio');
+                            await bot.sendAudio(chatId, output, { caption: `Audio uploaded. @tg_multitask_bot` });
+                        }
+
+                        fs.unlinkSync(output);  // Delete the file from the server
+                    }
+                });
             }
-
-            fs.unlinkSync(output);  // Faylni serverdan o'chirish
         } catch (error) {
-            bot.sendMessage(chatId, `Videoni yuklashda xatolik yuz berdi: ${error.message}. Iltimos, qaytadan urinib ko'ring.`);
+            bot.sendMessage(chatId, `Error downloading the video: ${error.message}. Please try again.`);
             console.error('Error during video download:', error);
         }
     } else {
-        bot.sendMessage(chatId, 'Yaroqsiz URL. Iltimos, to\'g\'ri YouTube yoki Instagram URL kiriting.');
+        bot.sendMessage(chatId, 'Invalid URL. Please enter a correct YouTube or Instagram URL.');
     }
 };
 
@@ -171,5 +228,5 @@ const extractVideoId = (url: string): string | null => {
     } catch (error) {
         console.error('Invalid URL:', error);
         return null;
-    }   
+    }
 };
